add DesiredBlockSize, which allows for "Unbounded" block sizes
this enables blocks to express that they'd like to take all available space.

this will become necessary as we begin to implement additional methods of layout, for instance a flex-list (or whatever we call it) can split it's available space evenly amongst its unbounded children. Previously, there would have been no way for the children to effectively indicate this, as they'd have to return a definite number, and that would make it difficult or impossible for the parent to allocate an appropriate space to each child.